refactor: Move every steps in a folder
