+++ src/cozy/react/components/InstallationPage/components/ConnectedStep.jsx
@@ -4,10 +4,10 @@ import NarrowContent from 'cozy-ui/transpiled/react/NarrowContent'
 import Stack from 'cozy-ui/transpiled/react/Stack'
 import Typography from 'cozy-ui/transpiled/react/Typography'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
-import { currentBrowser } from '../currentBrowser'
+import { currentBrowser } from '../../../currentBrowser'
 import chromeIllustration from 'cozy/react/assets/extension-connected-chrome.svg'
 import firefoxIllustration from 'cozy/react/assets/extension-connected-firefox.svg'
-import Help from './Help'
+import Help from '../../Help'
 import { ButtonLink } from 'cozy-ui/transpiled/react/Button'
 
 const illustrations = {
