+++ src/cozy/react/components/InstallationPage/components/InstallationStep.jsx
