+++ src/cozy/react/components/InstallationPage/components/InstalledStep.jsx
