+++ src/cozy/react/components/InstallationPage/components/SecurityStep.jsx
