+++ src/cozy/react/components/InstallationPage/components/styles.css
